+++ frontend/src/main.js
@@ -1,0 +1,11 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+import 'bootstrap/dist/css/bootstrap.css'
+import 'bootstrap-vue/dist/bootstrap-vue.css'
+import ToastPlugin from 'vue-toast-notification';
+
+import 'vue-toast-notification/dist/theme-bootstrap.css'
+
+const app = createApp(App)
+app.use(ToastPlugin);
+app.mount('#app');